+++ src/MissionChief_Map_Command_Toolkit.user.js
@@ -1,7 +1,7 @@
 // ==UserScript==
 // @name         MissionChief Map Command Toolkit
 // @namespace    https://github.com/Conroy1988/missionchief-map-command-toolkit
-// @version      8.2.4
+// @version      8.2.5
 // @description  MissionChief operational map command centre.
 // @author       Conroy1988
 // @license      MIT
@@ -453,7 +453,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND.
 
     const SCRIPT = {
         name: 'MissionChief Map Command Toolkit',
-        version: '8.2.4',
+        version: '8.2.5',
         author: 'Conroy1988',
         controlId: 'mc-map-command-toolkit-control',
         panelId: 'mc-map-command-toolkit-panel',
@@ -12902,12 +12902,42 @@ ${CUSTOM_VEHICLE_BADGE_SELECTOR}[data-mcms-theme="godfather"]{border-color:rgba(
     const TRANSPORT_SWEEP_OPTIONAL_RELEASE_SETTLE_MS = 6000;
     const TRANSPORT_SWEEP_OPTIONAL_RELEASE_REQUEST_TIMEOUT_MS = 12000;
 
+    function transportSweepOptionalReleaseTextWithBreaks(node) {
+        if (!node) return '';
+        const parts = [];
+        const walk = current => {
+            if (!current) return;
+            if (current.nodeType === 3) {
+                parts.push(String(current.nodeValue || ''));
+                return;
+            }
+            if (String(current.nodeName || '').toUpperCase() === 'BR') {
+                parts.push(' ');
+                return;
+            }
+            let children = [];
+            try { children = Array.from(current.childNodes || []); } catch (error) {}
+            children.forEach(walk);
+        };
+        walk(node);
+        return parts.join('').replace(/\s+/gu, ' ').trim();
+    }
+
     function transportSweepOptionalReleasePatientCountFromRow(row) {
-        const vehicleCell = row?.querySelector?.('td:first-child') || row;
-        const text = String(vehicleCell?.textContent || '').replace(/\s+/gu, ' ').trim();
+        if (!row) return null;
+        let patientCells = [];
+        try { patientCells = Array.from(row.querySelectorAll?.('td') || []); } catch (error) {}
+        const patientTextSource = patientCells.find(cell =>
+            /\bpatients?\s*:/iu.test(transportSweepOptionalReleaseTextWithBreaks(cell))
+        ) || row;
+        const text = transportSweepOptionalReleaseTextWithBreaks(patientTextSource);
         const match = text.match(/\bpatients?\s*:\s*(.+)$/iu);
         if (!match?.[1]) return null;
-        const names = match[1]
+        const patientText = match[1]
+            .replace(/\s+Release patient \(No reward\).*$/iu, '')
+            .replace(/\s*\([^()]*\)\s*$/u, '')
+            .trim();
+        const names = patientText
             .split(/\s*,\s*/u)
             .map(value => value.trim())
             .filter(Boolean);
